# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments/post/{postId}": {
      "get": {
        "description": "특정 게시글에 달린 모든 댓글을 조회합니다.",
        "operationId": "CommentsController_findAllByPostId",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 목록 반환",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "게시글의 모든 댓글 조회",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "description": "게시글에 댓글을 작성합니다. 로그인이 필요합니다.",
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "게시글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "댓글 작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "댓글을 생성할 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 작성",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{id}": {
      "patch": {
        "description": "자신이 작성한 댓글을 수정합니다. 로그인이 필요합니다.",
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "댓글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "댓글 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "댓글을 수정할 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "댓글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 수정",
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "description": "자신이 작성한 댓글을 삭제합니다. 로그인이 필요합니다.",
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "댓글 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "댓글 삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "댓글을 삭제할 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증에 실패했습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "권한이 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "댓글을 찾을 수 없습니다",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 삭제",
        "tags": [
          "comments"
        ]
      }
    }
  },
  "info": {
    "title": "COCOMOMO API",
    "description": "COCOMOMO 백엔드 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "에러 코드",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "에러 메시지",
            "example": "잘못된 요청입니다"
          },
          "error": {
            "type": "string",
            "description": "에러 타입",
            "example": "BadRequestException"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "댓글 ID (자동생성)",
            "example": 1
          },
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "좋은 게시글이네요!"
          },
          "userId": {
            "type": "number",
            "description": "작성자 ID",
            "example": 1
          },
          "postId": {
            "type": "number",
            "description": "게시글 ID",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "댓글 생성 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "댓글 수정 시간",
            "example": "2025-06-09T13:30:51+09:00"
          },
          "user": {
            "type": "object",
            "description": "작성자 정보",
            "example": {
              "id": 1,
              "username": "사용자"
            }
          }
        },
        "required": [
          "id",
          "content",
          "userId",
          "postId",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "좋은 글이네요!",
            "maxLength": 1000
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "댓글 내용",
            "example": "수정된 댓글입니다.",
            "maxLength": 1000
          }
        },
        "required": [
          "content"
        ]
      },
      "DeleteCommentResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "삭제 성공 여부",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      }
    }
  }
}